test: finish clearing test-code clippy findings (issue #818)
Fix the leftover non-doc clippy findings in test code after the suite
prune, including the last too_many_lines site:

- make_app_stub now builds through the existing `App::build` fixture
  (make_built_app) and overrides only the stub-specific defaults, instead
  of re-listing every App field.
- Type the daemon-route connect test seam as an explicit outcome enum
  (`DaemonRouteConnectOutcome`) so the always-Ok test overrides no longer
  trip `unnecessary_wraps`; `connect_daemon_route_endpoint` adapts with
  `into_result`.
- Replace float->int cast arithmetic in podcast/daemon-bootstrap tests
  with exact integer math, drop a non-item doc comment, hoist a nested
  test helper, switch a one-arm match to `if let`, and pass routing-matrix
  helper args by reference.

+++ src/app/tests/daemon_bootstrap.rs
@@ -98,7 +98,7 @@ fn local_daemon_app_keeps_live_abs_queue_and_reconciles_browse_on_adoption() {
 
     // Inject the live ABS queue slot (simulates the daemon broadcasting its
     // queue to the newly attached client via PlayerEvent::UnifiedQueueUpdated).
-    let acknowledged_position_ticks = (30.0 * mbv_core::api::TICKS_PER_SECOND as f64) as i64;
+    let acknowledged_position_ticks = 30 * mbv_core::api::TICKS_PER_SECOND;
     let abs_item = mbv_core::playback_queue::QueueItem::Audiobookshelf(
         mbv_core::playback_queue::AudiobookshelfQueueItem {
             library_item_id: "show-a".into(),
